new outgoing version
